fix: split UserEntryForm component with use of props
