+++ src/pages/forgot-password.js
@@ -1,13 +1,32 @@
-import React, { useEffect } from "react";
+import React, { useState, useEffect } from "react";
 import { Redirect } from "react-router-dom";
-import { useSelector } from "react-redux";
+import { useSelector, useDispatch } from "react-redux";
 import { getCookie } from "../utils/cookie";
+
+import { resetMyPassword } from "../services/reducers/password";
+
+import { Input } from "@ya.praktikum/react-developer-burger-ui-components";
 
 import UserEntryForm from "../components/UserEntryForm/UserEntryForm";
 
 function ForgotPasswordPage() {
+  const dispatch = useDispatch();
   const { accessToken } = useSelector((store) => store.auth);
   const token = getCookie("accessToken");
+
+  const [form, setForm] = useState({
+    email: "",
+  });
+
+  function onChange(e) {
+    setForm({ ...form, [e.target.name]: e.target.value });
+  }
+
+  function handleSubmitClick(e) {
+    e.preventDefault();
+
+    dispatch(resetMyPassword(form.email));
+  }
 
   useEffect(() => {}, [accessToken]);
 
@@ -15,7 +34,26 @@ function ForgotPasswordPage() {
     return <Redirect to={{ pathname: "/" }} />;
   }
 
-  return <UserEntryForm />;
+  return (
+    <UserEntryForm
+      onSubmit={handleSubmitClick}
+      title="Восстановление пароля"
+      buttonText="Восстановить"
+    >
+      <div className="mb-6">
+        <Input
+          type={"email"}
+          placeholder={"Укажите e-mail"}
+          onChange={onChange}
+          value={form.email}
+          name={"email"}
+          error={false}
+          errorText={"Ошибка"}
+          size={"default"}
+        />
+      </div>
+    </UserEntryForm>
+  );
 }
 
 export default ForgotPasswordPage;
